airbnb/requirements.txt
pyfarmhash==0.2.2
pandas==1.2.4
icecream==2.1.0
click==7.1.2
matplotlib==3.4.2
numpy==1.20.3
python-dotenv==0.17.1
scikit_learn==0.24.2
